Make $config available for language checks
Fixes errors such as:
PHP Notice:  Trying to get property 'LANGUAGE_PRIVATE' of non-object in /var/www/html/core/bot/userlanguage.php on line 6

While not needed in the main body of the function, the included
userlanguage.php does need it and is not a function of it's own but
running in the context of the top level function. -_-

I can't believe how long this issue has lurked.